Fix the build to support the new JsonDoubleValue.

+++ common/web/Json.cpp
@@ -143,7 +143,7 @@ void JsonObject::Add(const std::string &key, int i) {
   STLReplaceAndDelete(&m_members, key, new JsonIntValue(i));
 }
 
-void JsonObject::Add(const std::string &key, long double d) {
+void JsonObject::Add(const std::string &key, double d) {
   STLReplaceAndDelete(&m_members, key, new JsonDoubleValue(d));
 }
